Fix confusing dialyzer warnings for is_record/2 with illegal records
In commit 1858cb81391d2bce29b4b7620574ca60128cebf7, erl_expand_records
started to optimize is_record/2 in guards by replacing it with
pattern matching (if possible).

Unfortunately, dialyzer will no longer see the code before the
optimization, so any warnings produced in code such as:

  case ExprNotProducingRecord#rec{} of
         X when is_record(X, rec, N) -> ...

will refer to the optimized code and not the source code,
which is confusing for the user.

Introduce the no_is_record_optimization option for turning off
the optimization and use it in dialyzer.

Reported-by: Kostis Sagonas