Add a --fullpath option to Dialyzer
This change adds a --fullpath option to Dialyzer, which makes the
warning messages contain the full path of the corresponding file.

Original patch submitted by Magnus Henoch (legoscia) on 15/9/2010
and cooked to death in the 'pu' branch all this time.

The patch was essentially correct and most of it has been used as
is, but there have been some changes to make the code slightly
prettier, avoid some code duplication, and add documentation to
dialyzer's doc files and to its help message.