Shut off a dialyzer unmatched return warning
Dialyzer does not like list comprehensions for which the result is not
assigned to some variable (maybe anonymous).  In this particular case,
the comprehension, containing only a filter but no generator, was used
to save some lines, but IMO this type of code, while "cute", is not so
easy to comprehend.