Correct some type declarations
The types of some record fields were wrong. This file uses some module
local format of race warnings and when the analysis is done transforms
this format into the format of general dialyzer warnings. This change
fixes the type declaration of the race_warning field of the #races{}
record which was wrong.  In addition, declared the type of the state
field in some other record that, long ago, could not be declared since
mutually recursive types were not supported by Dialyzer's analysis.